+++ ext/semian/tickets.c
@@ -1,4 +1,4 @@
-#include <semian_tickets.h>
+#include <tickets.h>
 
 // "Private" function forward declarations
 static void
